bfasbot/requirements.txt
requests_futures
pyrogram
tgcrypto
python-dotenv
speedtest-cli 
img2pdf
inflect
timeago